Garbage_Sorter/requirements.txt
absl-py
annotated-types
anyio
astunparse
cachetools
certifi
charset-normalizer
click
colorama
distlib
dnspython
email-validator
fastapi
fastapi-cli
filelock
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h11
h5py
httpcore
httptools
httpx
idna
Jinja2
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
namex
numpy
oauthlib
opt-einsum
optree
orjson
packaging
Pillow
pip
platformdirs
protobuf
pyasn1
pyasn1-modules
pydantic
pydantic-core
Pygments
python-dotenv
python-multipart
PyYAML
requests
requests-oauthlib
rich
rsa
setuptools
shellingham
six
sniffio
starlette
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-hub
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
tf-keras
typer
typing-extensions
ujson
urllib3
uvicorn
virtualenv
watchfiles
websockets
Werkzeug
wheel
wrapt
tensorflow[and-cuda]
opencv-python